fix hex mouseover identification
- Created a transparent hex sprite so that it is always available as a
basic hex outline.  This sprite should not be used anywhere, but
duplicated if you need to create a new hex sprite using just a hex
outline.
- Filled-in the image0 of the hex sprite, to aid in precise collision
detection.  Prior to doing this, the system was not detecting a
collision if the mouse moved quickly to the interior of a hex
(presumably because it was transparent).
- changed the hex sprite collision detection to "precise". This deals
with misidentification due to overlapping collision masks when they were
rectangular.